Refactor: separate PhotoMarker component; separate map styles
